requirements.txt
streamlit>=1.24.0
psutil>=5.9.0
plyer>=2.1.0
pywin32>=305; platform_system == "Windows"
pandas>=1.5.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
sentence-transformers>=2.2.0
chromadb>=0.4.0
tenacity>=8.2.0 